fix(autoRequire): cache loaded implentation against name
An implementation file may be set by calling rely(name, file)
where file is a string. The first call to rely(name) should load
the file and store the module against name to prevent it being
reloaded on subsequent calls. This commit fixes bug that meant
the implementation was stored against the file name instead of
the module name.